requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.13
PyYAML>=6.0
joblib>=1.3
torch>=2.0
tqdm>=4.66
ipykernel>=6.25
nbformat>=5.9
nbconvert>=7.8
pytest>=7.4
